amended bnn script to do evaluation properly using 25 prediction samples debug

+++ bayesian_modeltrain.py
@@ -122,7 +122,7 @@ def evaluate_regression(model, valid_iterator, val_set, loss_fn, n_samples=25,  
                                      sample_nbr=3)
     # get a bunch of predictions to compare!
     print("generating %i predictions..."%n_samples)
-    preds = [model(ins) for i in range(n_samples)]
+    preds = [model(ins).float() for i in range(n_samples)]
     preds = torch.stach(preds)
     means = preds.mean(axis=0)
     stds = preds.std(axis=0)
